% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paste_with_space.R
\name{paste_with_space}
\alias{paste_with_space}
\title{Paste 2 strings with space}
\usage{
paste_with_space(x = "Meu", y = "Nome")
}
\arguments{
\item{x}{uma string}

\item{y}{outra string}
}
\description{
Paste 2 strings with space
}
